requirements.txt
Flask==1.1.1
gunicorn==20.0.4
Jinja2==2.10.3
psycopg2-binary==2.8.6
